Add terrain generation
Add terrain generation via a working diamond-square algorithm.
Still needs some documentation.
Written in such a way that each step of the algorithm is its own method.
This should *in theory* make it easier to increase/decrease map
resolution later on, if we decide to add that. It may, however, not be
the fastest way of performing the algorithm. It takes 1+ minute to
generate a terrain when N = 8, which may be an issue. I think a culprit
is that a lot of Vector2Ints are being created as the algorithm runs.
May be good to optimize this.