Add optimized version of diamond-square algorithm
Add a far faster version of the diamond-square algorithm. This version
is far more iteration-based, uses less objects, does not utilize many
lists, and is sectioned off so that it only returns a float[,] object.
After the heights have all been calculated, the vertices are *then*
calculated and the mesh is created.
Handles N=8 resolution within seconds.