Add the ability to toggle the cursor
When Escape is pressed, the mouse no longer rotates the camera and
instead moves a cursor on the screen. This allows the use of UI
elements that may be added in the future.
Also added a hint and in the bottom right to explain that Escape can
toggle the UI. This hint is toggled along with the UI.