Add incomplete day/night cycle
Implements a day/night cycle with two light sources - a sun and a moon.
On default, it goes at real time, though this can currently be changed
via the public fields in the DayNightCycle script. I've played around a
lot with the lighting values so it looks decent during both day and
night.
The reason why this is still incomplete is because, as demonstrated by
the scene on load, some unwanted shadows appear where the sun, when on
the horizon, can shine through the terrain, shadowing it with jaggies.
See for yourself. This will need a fix, or the day/night cycle will have
to be scrapped, unfortunately.