Alter GenerateTerrain to set water transform
Now passes in just the transform instead of the whole game object. Also,
instead of setting just the height, also sets its x and z coordinates to
match the center of the terrain and scales the plane properly as well.
Also, remove heights from the parameters since it's a field.