Fix Generate New Terrain button
The button now works again. Involved making several changes:
Modified TerrainColor to not have either a Start() or Update() method.
Instead, TerrainColor is assumed to exist by GenerateTerrain, which gets
a reference to it as a component and runs its methods as required.
Modified the structure of GenerateTerrain, moving most of the code out
of Start() and into two new methods - Initialize() and Generate(). See
comments for explanations.
This allows simplified button code as well, as all it now has to do is
call Generate() on the script, not worrying about deleting the old
terrain object and resetting a bunch of script values.