Constrain the player camera to the terrain
The camera can no longer move outside of the terrain. If the camera is
close to the terrain and the Generate New Terrain button, the camera
will be teleported above the new terrain if it is determined that it
might be generated above the camera.
Contains a few other minor changes.