Add non-functional UI elements
... for terrain generation parameters. This includes min/max corner
heights, min/max noise addition, noise reduction, terrain side length,
and n. It took a bit of effort to fit everything onto the left side of
the screen, so the "Generate New Terrain" button is smaller, but now
green to compensate to make it still stand out. None of the fields do
anything, as the scripts have not been touched.